Add files via upload
Updated files with successful add/view/del metadata of image url given. There is also more images give, I did not try some of them, they are fresh photos! Tried downloading new photos to try out. I did not return a statement therefore we were getting the strange error, it is now fixed. The issue we were having with most of these open-manipulate/view of files... I did not include a close() call. Additionally, it was realized the return statements for a few of the functions did not return the JSON data, which is why we were seeing something else within Postman. This has also been fixed. Things are working, yay.